Refactor CAT spends
